% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.deploy-package.r
\docType{package}
\name{glm.deploy}
\alias{glm.deploy}
\alias{glm.deploy-package}
\title{\packageTitle{glm.deploy}}
\description{
\packageDescription{glm.deploy}
}
\details{
All numeric variables used as input to the glm object are treated as doubles, whereas factors variables are treated as strings.\cr \cr
The \code{glm.deploy} package is compatible with the following link functions in the fitted glm object:
\itemize{
  \item identity
  \item probit
  \item cloglog
  \item log
  \item sqrt
  \item inverse
  \item 1/mu²
  \item logit
}
}
\seealso{
Functions:
\itemize{
  \item \code{\link{glm2c}}
  \item \code{\link{glm2java}}
}
URL: \url{https://github.com/oscarcastrolopez/glm.deploy}
}
\author{
\itemize{
  \item Oscar J. Castro-Lopez \email{oscarcastro@uas.edu.mx}
  \item Ines F. Vega-Lopez
}
}
