% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.deploy-package.r, R/glm.deploy.R
\docType{package}
\name{glm.deploy}
\alias{glm.deploy}
\alias{glm.deploy-package}
\alias{glm.deploy}
\title{glm.deploy.}
\usage{
glm.deploy(model, filename = NULL, language)
}
\description{
The glm.deploy models allows to export the specific scoring algorithm (predict) of a trained model in C and JAVA.Allowing to deploy trained gerealized linear models (glm) outside R. It generates two functions: 1) the equivalent to predict(type="response") and 2) the equivalent to predict(type="link").

the glm.deploy package allows to generate source code from glm objects to deploy/operationalize the scoring/predict functions outside R.
glm.deploy is used to generate the scoring functions of a trained GLM model in C or JAVA.
It should not be called directly
}
