% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.deploy.R
\name{glm2java}
\alias{glm2java}
\title{glm to java source code to deploy/operationalize outside R}
\usage{
glm2java(model, filename = NULL)
}
\arguments{
\item{model}{A trained object of class "glm".}

\item{filename}{The name of the file to save, if NULL the default is "glm_" plus the (intercept) variable name.}
}
\description{
glm2java is used to generate the scoring functions of a trained GLM model in the Java language.\cr
It generates a class with two functions:\cr
1) The equivalent to predict(type="response").\cr
2) The equivalent to predict(type="link").
}
\examples{
glm2java(glm(Y ~ ., family = binomial(logit), data=data), "MyFileName")
}
\author{
Oscar J. Castro-Lopez, Ines F. Vega-Lopez
}
